Avoid circular dependency issue
