+++ tests/unit/amo/utils/test_index.js
@@ -1,33 +1,8 @@
-import NotAuthorized from 'amo/components/ErrorPage/NotAuthorized';
-import NotFound from 'amo/components/ErrorPage/NotFound';
-import ServerError from 'amo/components/ErrorPage/ServerError';
-import {
-  getCanonicalURL,
-  getErrorComponent,
-  shouldShowThemes,
-} from 'amo/utils';
+import { getCanonicalURL, shouldShowThemes } from 'amo/utils';
 import { CLIENT_APP_ANDROID, CLIENT_APP_FIREFOX } from 'core/constants';
 import { getFakeConfig } from 'tests/unit/helpers';
 
 describe(__filename, () => {
-  describe('getErrorComponent', () => {
-    it('returns a NotAuthorized component for 401 errors', () => {
-      expect(getErrorComponent(401)).toEqual(NotAuthorized);
-    });
-
-    it('returns a NotFound component for 404 errors', () => {
-      expect(getErrorComponent(404)).toEqual(NotFound);
-    });
-
-    it('returns a ServerError component for 500 errors', () => {
-      expect(getErrorComponent(500)).toEqual(ServerError);
-    });
-
-    it('returns a ServerError component by default', () => {
-      expect(getErrorComponent(501)).toEqual(ServerError);
-    });
-  });
-
   describe('getCanonicalURL', () => {
     it(`returns an absolute canonical URL`, () => {
       const locationPathname = '/path/name';
